backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
opencv-python==4.8.0.76
numpy==1.26.0
pillow==10.0.1